improve memory usage for hostTrackData (#431)
This PR improves the data usage of the HostTrackData.

Before, when tracks were deleted, the g4idToGPUid map could not be
cleaned as it is still needed if a track goes from GPU to CPU to GPU. In
that case, we have to ensure that the second time the track goes to the
GPU it has the same track ID as the first time, as the track ID is used
to seed the RNG. However, when the track is killed, we can safely delete
all of it. Now, we have two functions, one to fully remove a track from
the hostTrackData when the track is killed and one to "retire" it to the
CPU so that we keep the g4idToGPUid in case the CPU decides to put the
track back on the GPU. This is only needed in case the of a
out-of-gpu-region leak.

Also, a destructor is added to clean all data, as before the data was
only cleaned upon a new event, leading to a memory leak