Fix Async State Machine (#432)
There was a subtle bug in the state machine in async AdePT which could
lead to some hits being processed by the wrong event, which would cause
crashes with the hostTrackData (as it was potentially not available for
that track, leading to crashes).

The issue happened as follows:

1. A thread requests a hit flush, which leads to swapping of the buffer,
copying of the hits, and putting them in the hitQueue of the workers.
2. In the meantime, a second thread requests a hitflush, but it cannot
swap the buffer as the hits are still being copied back from the first
step.
3. Now, the transfer is finished and we advance to the block right after
the notifying of the hit management thread, we call this:
```c++
      // *** Notify G4 workers if their events completed ***
      if (std::any_of(eventStates.begin(), eventStates.end(), [](const std::atomic<EventState> &state) {
            return state.load(std::memory_order_acquire) == EventState::DeviceFlushed;
          })) {
        // Notify HitProcessingThread to notify the workers. Do not notify workers directly, as this could bypass the
        // processing of hits
        hitProcessing->cv.notify_one();
      }
```
If now a third thread is already in the state of `DeviceFlushed`, the
hit management loop is started. In there, unconditionally, all
`RequestHitFlush` states are advanced to `FlushingHits` and subsequently
to `HitsFlushed` even if the swap was not executed yet, so that the
second thread jumps ahead of the state machine although its GPU steps
are not copied back yet, as the buffer couldn't be swapped. Then, they
can end up in the next event as the state machine can run through.

This is fixed by adding another state, `HitBufferSwapped`. This guards
against the case that a thread can advance from `RequestHitFlush` to
`FlushingHits` and beyond despite its hits not being flushed.

Furthermore in this PR the injection bug is handled a bit more
gracefully by resetting the state. It has not been observed again, this
is rather a precaution.


Validated in:

- [x] example1 with testEm3
- [x] Gauss
- [x] Athena