Publish CMake variables, revert unsafe option (#434)
This PR fixes an issue observed in Athena:

Previously, AdePT did not publish the CMake variable such as
`ASYNC_MODE` etc to the UserCode. Then, while AdePT was compiled
correctly as an external, when parts of it where used in the physics
list, the variables were not set correctly.

This is now done correctly.

In the process, it was uncovered that the retiring of the HostTrackData
for tracks that leave the GPU region is not safe, as it is not
guaranteed that the GPU Steps are processed before the leaked tracks.
This lead to secondaries missing the link to their parents, as the
parents' hostTrackData was deleted when the initializing step of the
secondary was processed.