Reuse already initialized HepEm tracks in split kernels mode (#433)
When using split kernels, we keep a pool of HepEm tracks that we use to
avoid instantiating and initializing a track in each kernel of a step.
We can take advantage of this by re-using and not re-initializing tracks
after the first step for a particle.

Acessing and setting HepEm track data is a large source of long
scoreboard stalls. This change provides a speedup of around 8% for the
split kernels. The same optimization has also been tried on the
monolithic kernels but it doesn't seem to give any speedup in that case.

There is a small diference in energy deposition, which is likely caused
by the fact that fields like the numIALeft or mscData are stored in
`double` by G4HepEm and `float` by AdePT. By keeping these fields in the
HepEm track and avoiding these two conversions the results are slightly
altered (`double` to `float` to store the values in the AdePT track and
`float` to `double` to initialise the HepEm track in the following
iteration).

As we don't need to store some of the values from the HepEm track, these
changes also allow to save some space in the tracks, which is much
needed for the split kernels, as we already need extra memory in the
form of extra track fields and the HepEm track pool.

```
RTX 4080, CMS2026 (Scoring disabled)
8 threads, 32 TTbar events
```
| Kernels             | Time |
|---------------------|------|
| Monolithic          | 256  |
| Split               | 260  |
| Split + HepEm reuse | 240  |

`ncu` shows the reduction in long scoreboard stalls for the
`ElectronHowFar` kernel:
**Master**
<img width="1143" height="47" alt="image"
src="https://github.com/user-attachments/assets/a1255357-c639-4ddb-9234-6a634c9a91d9"
/>
<img width="2366" height="191" alt="image"
src="https://github.com/user-attachments/assets/bf227ac9-574c-4118-af56-8bfd3553a20b"
/>

**Reuse HepEm**
<img width="1143" height="47" alt="image"
src="https://github.com/user-attachments/assets/37bc081d-8e72-4965-a3ae-622115015659"
/>
<img width="2366" height="191" alt="image"
src="https://github.com/user-attachments/assets/3ea12d34-32ac-4789-b2a1-67504e1554b3"
/>