Improve HostTrackDataMapper (#437)
This PR improves the handling of the hostTrackData in the following way:

Before, the `retiringToCPU` was deactivated as it was assumed to be
unsafe. The problem was in the `AdePTTrackingManager`:

When a track is given back to the CPU for tracking, the hostTrackData
can be deleted, except for the G4-To-GPU ID map because for
reproducibility, if the track goes back to the GPU, it would require to
get the same GPU ID.
Previously, it was missing to re-instate all of the HostTrackData in
case the G4-to-GPU ID entry was found. Then, while having the correct
GPU ID, the remaining hostTrackData was zero, leading to crashes when
the secondaries were generated by the particle that went from GPU to CPU
back to GPU.

This is now fixed, and the hostTrackData is also set when only the GPU
id was recovered.

Furthermore, unit tests were added for the hostTrackData to test the
core functionality.