Async AdePT initialization cleanup (#446)
Due to the nuclear processes, we need each integration layer object to
hold a reference to the thread's HepEmTrackingManager. This was achieved
by initializing each integration object per thread after the AdePT
common initialization. However, the code gave the impression that this
was happening in the AsyncAdePTTransport constructor, while in reality
this is called only once per run, by the first thread that runs the
initialization.

With these changes, the unneeded code is removed from the constructor,
and the integration objects are default constructed, rather than being
created at AdePT initialization. Each thread stores its
HepEmTrackingManager on the corresponding integration object with a call
after retrieving the pointer to the common AsyncAdePTTransport.