Changing vecgeom::Precision to double (#447)
Drop the usage of VECGEOM_FLOAT_PRECISION (Precision=float) and replace
Precision<->double.

This prepares for the `v2` release of VecGeom where `Precision` lives
only in `vecgeom` namespace and will break the compilation of AdePT. The
change of Precision->double will preserve the capability to run also
with older VecGeom. VECGEOM_FLOAT_PRECISION was an experimental feature
known to have big limitations for realistic setups anyway.