Consumption benchmarking (#443)
Adds the `ENABLE_POWER_METER` cmake option in order to do power
consumption measurements.

If the option is enabled, we use the [power
meter](https://github.com/apt-sim/power_meter) library to launch a
background thread taking measurements from CPU and GPU. The monitoring
is started once AdePT initialization is finished, and ends when shutting
down the application.

Note that if the option is enabled, simulations need to be run with root
access, which is required in order to read the CPU's MSR registers.

Prior to Geant4 11.3 the specialized tracking managers were not deleted
by Geant4. Since the power meter is stopped in the tracking manager
destructor, this will lead to an error message in older geant4 versions
when the program ends without joining the power meter thread. This
should not cause a memory leak, since `terminate()` will be called on
the unjoined thread.