[mini] fix printout of track id, reduce printout of tracking manager (#449)
This PR addresses two issues:

1. the format for printing the track id was `%ld` but should be `%lu` as
the track ids are unsigned long ints. This PR fixes the issue, which
will prevent output like this:
```
AtlasG4Tf 10:10:23 Killing e-/+ event 20390001 track -781294641970619956 E=0.758568 lvol=62698 after 90 steps with zeroStepCounter 26
AtlasG4Tf 10:10:23 Killing e-/+ event 20390001 track -9060344117526303898 E=2.369207 lvol=63312 after 38 steps with zeroStepCounter 26
```
to always print the correct output.

2. the fVerbosity requirement in the AdePTTrackingManager is increased a
bit. Note that this is not the verbosity used in the AdePT UI commands,
but the one that is given to the tracking manager when created. That
corresponds typically to the verbosity of the physics list. Maybe this
printout should indeed be linked to the AdePT verbosity command, but for
now I just increased it to at least verbosity 2, as verbosity 1 is often
used in the physics list (e.g., in CMSSW).