make speed of light region-dependent (#451)
Previously, the SpeedOfLight option would just remove all EM physics.
However, it is interesting to see the fraction of work that is offloaded
to the GPU.
By moving the SpeedOfLight check to where the track would be put on the
GPU, a region-specific analysis can be made. For example, when not
offloading certain regions to GPU in ATLAS, with this feature the
speed-of-light of that setup can be assessed, at the moment only the
total EM fraction can be assessed.