Add extendable, experiment-specific SteppingAction (#453)
This PR adds extendable, experiment-specific SteppingActions.

They can be enabled via `-DADEPT_STEPPINGACTION=LHCb|CMS`. So far, only
the SteppingAction for LHCb and CMS are implemented, but further actions
could be added.

A struct with parameters for the stepping action is passed to the kernel
(it is specific to the experiment), and the kernels are templated over
the SteppingAction.
Then, the kernel calls the correct stepping action.

No run time difference was observed in the LHCb minbias runs.

Note that so far this is only for the monolithic kernels, the split
kernels would first need a setup PR similar to #429