unify the usage of the VolAuxArray between sync and async AdePT (#454)
This PR unifies the usage of the VolAuxArray between the sync and async
kernels in AdePT:

While available as a global in sync AdePT, it was passed as an argument,
while in async AdePT, the global instance was used.

Since this is truly global data like geometry and the g4hepem data, it
is fine to just use the global instance. This PR unifies the usage and
cleans the kernels.

Additionally, a few outdated `#ifdef ADEPT_USE_SURF` were removed, which
were needed before the surface model was merged.