[mini] Store G4RegionId  (#455)
This PR changes the fGPURegion to actually store the G4RegionId. Before,
it was 0 for non-GPU and 1 for GPU (should have been a bool), now it
actually stores the G4RegionId if it is used on the GPU, otherwise it is
set to -1. This is in preparation for Woodcock tracking on the GPU.