[mini] move position when step size advance fails (#450)
This mini PR fixes a bug reported by code analysis from @dkonst13:

The the loop of reducing the step size when stuck on the boundary, the
`move` was adjusted, but the position was not changed. As the `move`
affects the remains and the stepDone, the position must be also adjusted
here, otherwise it is not consistent.