unify passing of queues and tracks to async kernels (#456)
This PR cleans the passing of arguments to the async kernels.
Before, multiple queues and the "secondaries" were passed, which
included the queues themselves. Then, slots were taken and passed to
queues directly etc.

This PR unifies the usage of the queues, now only the `ParticleManager`
needs to be passed, which holds the `SpeciesParticleManager`, which was
the previous `ParticleGenerator`. It is extended by the few missing
queues, and has new function to e.g., directly copy a track of a
corresponding slot to the leaks, which simplifies the kernels.

This simplifies the arguments of kernel launches as well.

This PR is in preparation for the Woodcock tracking.