[mini] clean FillFromDeviceBuffer (#459)
This PR cleans `FillFromDeviceBuffer` kernel:

When looping over the particles to be copied back, the index starts at
`nAlreadyTransfered` and was going to `nAlreadyTransferred +
maxFromDeviceBuffer`. Via an if statement, out of bounds access was
prevented.
This PR proposes to directly fix the range correctly in the loop,
enabling to remove the out of bounds check later. Thus, this is some
minor cleaning PR

Found via code analysis by @dkonst13.