Fix bug in stalling of the GPUs when running out of hitslots (#460)
This fixes a bug in the stalling of the GPU in case the hitBuffer is
overflowing. Since the calculation of `nextStepMightFail` is based on
unsigned int parameters, the subtraction was causing false negatives in
case of wraparounds if `hitBufferSafetyFactor * maxInFlight ` was larger
than ` gpuState.fHitScoring->HitCapacity() / numThread`. By changing the
condition to an addition
```
        bool nextStepMightFail = gpuState.stats->hitBufferOccupancy + hitBufferSafetyFactor * maxInFlight >=
                                 gpuState.fHitScoring->HitCapacity() / numThreads;
```
this is safely resolved.

Furthermore, as a track can cause multiple (3) hits per track, (one for
each secondary, one for the energy it deposits itself), the previous
hardcoded hitBufferSafetyFactor of 1.5 was dangerous, so now it is an UI
parameter. It doesn't make sense to hardcode a higher (safe) value, as
it would slow down the settings where a lower factor is safe.

Furthermore, two printouts that are per track are increased to a higher
verbosity, as those render larger simulations useless, but other
printouts (e.g., whether the hitbuffer is copied) were at the same
level.