Fix setting of G4Touchable when giving track back to G4 (#461)
This small PR fixes the handling of the G4Touchable when tracks are
returned to G4.
As the UserStackingAction might check on the Touchable or NextTouchable,
they must be set before giving it back to G4. This caused a crash in
CMSSW.

At the same time, previously, every step was located in G4 when given to
AdePT via `SetInitialStep` (see
https://gitlab.cern.ch/geant4/geant4/-/blob/master/source/tracking/src/G4SteppingManager.cc?ref_type=heads#L309-L326).
Now, this is called only for initial Steps (`step number == 0`).

This change slightly improves performance for offloading only the HCAL
and ECAL in example1 using the CMS2018 geometry.

Also a minor cleaning was done: the `FillG4NavigationHistory` now
receives a const reference instead of by value.