Fix synchronization of injection (#445)
This PR fixes the infamous "injection bug", a small race condition that
could lead to advancing the state machine too fast.
Note that this bug was harmless as it was checked for and case it was
detected, the state machine was simply reset back to transport.

This PR cleans the injection and fixes the bug:

1. Previously, the kernel of the enqueing of the tracks was
[synchronized with the
iteration](https://github.com/apt-sim/AdePT/pull/445/files#diff-e7786b5f8e89172d219a16625b438e59af0580796e9f7c19de3a946cf65e8085L1575-L1580)
and then the state machine was [advanced in the next
iteration](https://github.com/apt-sim/AdePT/pull/445/files#diff-e7786b5f8e89172d219a16625b438e59af0580796e9f7c19de3a946cf65e8085L1075-L1081).
This PR removes this unnecessary synchronization, by using host function
callbacks that advance the state machine to `InjectionComplete` as soon
as the kernel finishes, exactly as it was already done before for the
`InjectState::ReadyToEnqueue` after the HtoD copy had finished. This
cleans the implementation and avoids synchronization by hand but doesn't
fix the bug.

2. The bug was happening at the fetching of slots for the G4 workers:
Between the G4 worker [taking current buffer and locking
it](https://github.com/apt-sim/AdePT/pull/445/files#diff-da9c52debb6803213b36f2cb01ba131a11aeaa9559e8fd189188ecf3a81b651dL272-L274),
the AdePT worker thread could [swap out the buffer
](https://github.com/apt-sim/AdePT/pull/445/files#diff-e7786b5f8e89172d219a16625b438e59af0580796e9f7c19de3a946cf65e8085L1086-L1087),
such that the G4 worker would increment and add a slot to the just
swapped out buffer. Then, the state machine could advance, while this
"lost slot" would then appear later. That's why the error only ever
showed a single track being tracked in the wrong state, as this would
happen very rarely and then only for a single track at the same time.
By re-checking that the buffer is still the same after the G4 worker
takes the lock (and just dropping it and taking the new active buffer if
it isn't), this slip is fixed. The bug fix was verified in an Athena
simulation where next to the calorimeters also the ITK was offloaded,
which would trigger the bug previously. Now, the bug disappeared having
the same run time.

3. Before, when there was contention in the ToDeviceBuffer, the G4
workers would hold the lock on the buffer and then go to sleep (and then
release it at the end of the iteration of the while loop before trying
again). Keeping the lock would prevent the AdePT worker thread to
actually swap out the buffer. By releasing the lock before going to
sleep, the AdePT worker thread can swap out the buffer faster, reducing
the stall time when the ToDeviceBuffer is full. This was tested in
example1 with ttbar events in CMS 2024 geometry. By artifically
increasing the pressure by reducing the ToDeviceBuffer, the behaviour
was tested:
For three runs, the run time in seconds was
master: 231.34, 223.851, 227.222
And with this PR: 190.964, 192.325, 191.497
Note that this was typically not performance critical. The current
Athena production runs give the same run time with and without this PR.
However, offloading more parts of the detector (also the ITKPixel and
ITKStrip), which increases the ToDevice and fromDevice exchanges of
tracks due to the small gap between the regions, a small performance
improvement was observed.

 
This PR does not change any physics result.